third_party/WebKit/LayoutTests/fast/events/event-fire-disconnected-bubbling-custom-event.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Custom events should propagate in disconnected DOM trees.");

var eventReceivedOnParent = false;
var eventReceivedOnChild = false;

var parent = document.createElement('div');
var child = document.createElement('div');
parent.appendChild(child);
parent.addEventListener('foo', function(e) { eventReceivedOnParent = true; } );
child.addEventListener('foo', function(e) { eventReceivedOnChild = true; } );
var event = document.createEvent('Event');
event.initEvent('foo', true, true);
child.dispatchEvent(event);

shouldBeTrue("eventReceivedOnParent");
shouldBeTrue("eventReceivedOnChild");
</script>
</body>
</html>
